pyproject.toml
[tool.poetry]
name = "fastapi-auth"
version = "0.1.0"
description = "An example of FastAPI backend with an authentication system"
authors = ["AliRzaev <alirzaev1996@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.70.0"
uvicorn = "^0.16.0"
alembic = "^1.7.5"
SQLAlchemy = "^1.4.28"
pydantic = {extras = ["email"], version = "^1.8.2"}
psycopg2-binary = "^2.9.2"
python-multipart = "^0.0.5"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
python-jose = {extras = ["cryptography"], version = "^3.1.0"}

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
